move dispatchGui to DispatchManager
