+++ src/main/java/com/github/manolo8/darkbot/core/manager/DispatchManager.java
@@ -4,6 +4,7 @@ import com.github.manolo8.darkbot.Main;
 import com.github.manolo8.darkbot.core.objects.Gui;
 import com.github.manolo8.darkbot.core.objects.facades.DispatchProxy;
 import com.github.manolo8.darkbot.core.objects.facades.DispatchRetrieverMediator;
+import com.github.manolo8.darkbot.utils.Time;
 import eu.darkbot.api.managers.DispatchAPI;
 
 import java.util.List;
@@ -40,6 +41,50 @@ public class DispatchManager extends Gui implements DispatchAPI {
 
     public Retriever getSelectedRetriever() {
         return mediator.getSelectedRetriever();
+    }
+
+    public boolean openRetrieverTab() {
+        if (show(true)) {
+            click(80, 70);
+        }
+        return isAnimationDone();
+    }
+
+    public boolean openAvailableTab() {
+        if (show(true)) {
+            click(80, 100);
+        }
+        return isAnimationDone();
+    }
+
+    public boolean clickFirstItem() {
+        if (show(true)) {
+            Time.sleep(25);
+            click(300, 150);
+        }
+        return isAnimationDone();
+    }
+
+    public boolean clickHire() {
+        if (show(true)) {
+            click(700, 375);
+        }
+        return isAnimationDone();
+    }
+
+    public boolean openInProgressTab() {
+        if (show(true)) {
+            click(200, 100);
+        }
+        return isAnimationDone();
+    }
+
+    public boolean clickCollect(int i) {
+        if (openInProgressTab()) {
+            Time.sleep(25);
+            click(260, 160 + (41 * i));
+        }
+        return isAnimationDone();
     }
 }
